test: make BiquadFilterChromium closer to original implementation using doubles

+++ packages/react-native-audio-api/common/cpp/audioapi/core/effects/BiquadFilterNode.h
@@ -66,7 +66,7 @@ class BiquadFilterNode : public AudioNode {
       const float *frequencyArray,
       float *magResponseOutput,
       float *phaseResponseOutput,
-      int length);
+      size_t length);
 
  protected:
   std::shared_ptr<AudioBus> processNode(
